fix: query health further in the past (#14)
fix: query health further in the past

From time to time health status returns wrong value which is not
returned for the same interval later. Assuming there might be some
concurrency in graphite causing data to be distorted (i.e. since we use
precentages and similar stuff portion of metrics can be updated
while other portion is still not written). To try to eliminate this
the reported is modified to query interval not -1m..now, but -2m..-1m

Reviewed-by: Vladimir Hasko <vladimirhasko@gmail.com>